[fix ##147903479] beforeEach for documents tests backend
