+++ server/tests/serverSide/testUser.js
@@ -4,7 +4,6 @@ const chai = require('chai');
 let should = chai.should();
 const chaiHttp = require('chai-http');
 const app = require('../../../app');
-const User = ('../../controllers/users.js');
 
 chai.use(chaiHttp);
 let token = '';
